Development/requirements.txt
# Perseus Data Pipeline Requirements

# Core dependencies
numpy>=1.21.0
scipy>=1.7.0
pyelftools>=0.32

# Graph processing
networkx>=2.6.0

# Plotting Features
matplotlib>=3.10.8

# Disassembly
capstone>=4.0.2

# Machine Learning
torch>=2.0.0
torch-geometric>=2.3.0
transformers>=4.36.0
peft>=0.7.0
bitsandbytes>=0.41.0
datasets>=2.14.0
accelerate>=0.25.0
trl>=0.7.0

# Data processing
pandas>=1.3.0

# Serialization
h5py>=3.0.0

# Utilities
tqdm>=4.62.0
pyyaml>=5.4.0

# Logging and monitoring
tensorboard>=2.8.0
wandb>=0.12.0  # Optional: for experiment tracking

# Code analysis (optional)
# pygraphviz>=1.9  # For CFG visualization (requires graphviz system package)

# Testing
pytest>=7.0.0
pytest-cov>=3.0.0

# Code quality
black>=22.0.0
flake8>=4.0.0
mypy>=0.950
